Add extra file for sodium
