+++ src/majordomo/CMakeLists.txt
@@ -3,6 +3,7 @@ add_library(majordomo INTERFACE
     include/majordomo/Broker.hpp
     include/majordomo/MockClient.hpp
     include/majordomo/Constants.hpp
+    include/majordomo/Cryptography.hpp
     include/majordomo/Rbac.hpp
     include/majordomo/RestBackend.hpp
     include/majordomo/Settings.hpp
